// lib/process-services/src/lib/form/widgets/content-widget/attach-file-widget-dialog-component.interface.ts
/*!
 * @license
 * Copyright © 2005-2023 Hyland Software, Inc. and its affiliates. All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Subject } from 'rxjs';
import { Node } from '@alfresco/js-api';

export interface AttachFileWidgetDialogComponentData {
    title: string;
    actionName?: string;
    selected: Subject<Node[]>;
    currentFolderId: string;
    ecmHost: string;
    context?: string;
    isSelectionValid?: (entry: Node) => boolean;
    showFilesInResult?: boolean;
    loginOnly?: boolean;
    accountIdentifier?: string;
    registerExternalHost?: (...args) => void;
}